// src/type.ts
/* export enum CATEGORY {
  ACCESSORIES = "Acessórios",
  ELECTRONICS = "Eletrônicos",
}

export type User = {
    id: string;
    email: string;
    password: string;
  };
  
  export type Product = {
    id: string;
    name: string;
    price: number;
    category: CATEGORY;
  };
  
  export type Purchase = {
    userId: string;
    productId: string;
    quantity: number;
    totalPrice: number;
  };
   */